refactor(services): split GitHub API services into separate files
- Move GitHub API services into individual files in the github directory
- Create new files for commits, issues, pullRequests, and utils
- Update import paths in components to use new module structure
- Add types file to define interfaces and types for GitHub API data
- Update githubGraphQLService.ts to re-export from new module structure
